+++ Libplanet.Tests/Action/ActionEvaluatorTest.cs
@@ -583,7 +583,8 @@ namespace Libplanet.Tests.Action
             var actionEvaluator = new ActionEvaluator<DumbAction>(
                 policyBlockAction: null,
                 blockChainStates: NullChainStates<DumbAction>.Instance,
-                trieGetter: null, genesisHash: null);
+                trieGetter: null,
+                genesisHash: tx.GenesisHash);
 
             foreach (bool rehearsal in new[] { false, true })
             {
@@ -710,7 +711,8 @@ namespace Libplanet.Tests.Action
             var actionEvaluator = new ActionEvaluator<ThrowException>(
                 policyBlockAction: null,
                 blockChainStates: NullChainStates<ThrowException>.Instance,
-                trieGetter: null, genesisHash: null);
+                trieGetter: null,
+                genesisHash: tx.GenesisHash);
             var block = new BlockContent<ThrowException>
             {
                 Index = 123,
@@ -748,6 +750,7 @@ namespace Libplanet.Tests.Action
 
             Block<Arithmetic> blockA = await fx.Mine();
             ActionEvaluation[] evalsA = ActionEvaluator<DumbAction>.EvaluateActions(
+                txA.GenesisHash,
                 blockA.PreEvaluationHash,
                 blockIndex: blockA.Index,
                 txid: txA.Id,
@@ -797,6 +800,7 @@ namespace Libplanet.Tests.Action
 
             Block<Arithmetic> blockB = await fx.Mine();
             ActionEvaluation[] evalsB = ActionEvaluator<DumbAction>.EvaluateActions(
+                txB.GenesisHash,
                 blockB.PreEvaluationHash,
                 blockIndex: blockB.Index,
                 txid: txB.Id,
@@ -1071,6 +1075,29 @@ namespace Libplanet.Tests.Action
             };
 
             return (addresses, txs);
+        }
+
+        [Fact]
+        private async void CheckGenesisHashInAction()
+        {
+            var chain = MakeBlockChain<EvaluateTestAction>(
+                    policy: new BlockPolicy<EvaluateTestAction>(),
+                    store: _storeFx.Store,
+                    stateStore: _storeFx.StateStore);
+            var privateKey = new PrivateKey();
+            var action = new EvaluateTestAction();
+            var tx = Transaction<EvaluateTestAction>.Create(
+                nonce: 0,
+                privateKey: privateKey,
+                genesisHash: chain.Genesis.Hash,
+                actions: new[] { action });
+            chain.StageTransaction(tx);
+            var miner = new PrivateKey();
+            await chain.MineBlock(miner);
+            var evaluations = chain.ActionEvaluator.Evaluate(
+                chain.Tip,
+                StateCompleterSet<EvaluateTestAction>.Recalculate);
+            Assert.Equal(chain.Genesis.Hash, evaluations[0].InputContext.GenesisHash);
         }
 
         private sealed class EvaluateTestAction : IAction