+++ CHANGES.md
@@ -26,6 +26,8 @@ To be released.
    - Added `BlockChain<T>.FindNextHashes(BlockLocator, BlockHash?, int)` method.
    - Added `BlockChain<T>.Fork(BlockHash, bool)` method.
    - Added `BlockChain<T>.GetBlockLocator(int)` method.
+ -  Added Genesis Hash property in `IActionContext`. [[#1972], [#2179]]
+   -  Added Genesis Hash property in `ActionEvaluator`.
 
 ### Behavioral changes
 
@@ -40,6 +42,7 @@ To be released.
 
 ### CLI tools
 
+
 [#1762]: https://github.com/planetarium/libplanet/issues/1762
 [#2047]: https://github.com/planetarium/libplanet/issues/2047
 [#2140]: https://github.com/planetarium/libplanet/pull/2140
@@ -48,6 +51,8 @@ To be released.
 [#2159]: https://github.com/planetarium/libplanet/pull/2159
 [#2162]: https://github.com/planetarium/libplanet/issues/2162
 [#2171]: https://github.com/planetarium/libplanet/pull/2171
+[#1972]: https://github.com/planetarium/libplanet/issues/1972
+[#2179]: https://github.com/planetarium/libplanet/pull/2179
 
 
 Version 0.39.0